ch4/ofi: refactor vci-related initialization code
Move code that are related to multiple-vci setup to ofi_vci.c.
