+++ src/mpid/ch4/netmod/ofi/ofi_init.h
@@ -32,8 +32,11 @@ void MPIDI_OFI_update_global_settings(struct fi_info *prov);
 /* Determine if NIC has already been included in others */
 bool MPIDI_OFI_nic_already_used(const struct fi_info *prov, struct fi_info **others, int nic_count);
 
+int MPIDI_OFI_create_vci_context(int vci, int nic);
 int MPIDI_OFI_addr_exchange_root_ctx(void);
 int MPIDI_OFI_addr_exchange_all_ctx(void);
+int MPIDI_OFI_am_init(int vci);
+int MPIDI_OFI_am_post_recv(int vci, int nic);
 
 bool MPIDI_OFI_nic_is_up(struct fi_info *prov);
